requirements.txt
Flask==2.3.3
flask-restx==1.2.0
google-api-python-client==2.85.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
fhir.resources==6.4.0        # optional, for FHIR object modeling
gunicorn==20.1.0
